Move BaseOpenGLRenderer into a separate module
This means we can extend the BaseOpenGLRenderer without having depend on PyOpenGL. There are several other wrappers out there.
